Moved the PassPacket function from routers.py to helper.py to maintain consistency.
