+++ routers.py
@@ -43,29 +43,6 @@ pathChanToAnn = pathAnnToChan[::-1]
 # Everything should be local, make sure all ports are under this IP
 localHost = helper.localHost
 
-# This function will inspect incoming packet and send it to the next relevant node
-def PassPacket(shortestPath, routerName, packetOnTheWay):
-    
-    # Identify which router the packet is at and send it to the next relevant
-    nextRouterIndex = shortestPath.index(routerName) + 1
-
-    # Make sure not to get index out of bounds
-    if nextRouterIndex < len(shortestPath):
-        nextRouterName = shortestPath[nextRouterIndex]
-        nextRouterPort = helper.namesAndPorts.get(nextRouterName)  
-
-        try:
-            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            
-            # Connect to server and send data
-            sock.connect((localHost, nextRouterPort))
-            sock.sendall(packetOnTheWay)
-        
-        except ConnectionRefusedError:
-            print(nextRouterName + " is offline.")
-
-        finally:
-            sock.close()
 
 # -----------------------------------------------------------------------------
 # This class can be instantiated to create a multithreaded server multithreaded 
@@ -92,27 +69,27 @@ def TCPHandler(routerName):
 
             # Packet is from Ann to Jan
             if sourceID == helper.namesAndPorts.get('Ann') and destinationID == helper.namesAndPorts.get('Jan'):
-                PassPacket(pathAnnToJan, routerName, packetOnTheWay)
+                helper.PassPacket(pathAnnToJan, routerName, packetOnTheWay)
             
             # Packet is from Jan to Ann
             elif sourceID == helper.namesAndPorts.get('Jan') and destinationID == helper.namesAndPorts.get('Ann'):    
-                PassPacket(pathJanToAnn, routerName, packetOnTheWay)
+                helper.PassPacket(pathJanToAnn, routerName, packetOnTheWay)
             
             # Packet is from Jan to Chan
             elif sourceID == helper.namesAndPorts.get('Jan') and destinationID == helper.namesAndPorts.get('Chan'):    
-                PassPacket(pathJanToChan, routerName, packetOnTheWay)
+                helper.PassPacket(pathJanToChan, routerName, packetOnTheWay)
 
             # Packet is from Chan to Jan
             elif sourceID == helper.namesAndPorts.get('Chan') and destinationID == helper.namesAndPorts.get('Jan'):    
-                PassPacket(pathChanToJan, routerName, packetOnTheWay)
+                helper.PassPacket(pathChanToJan, routerName, packetOnTheWay)
 
             # Packet is from Ann to Chan
             elif sourceID == helper.namesAndPorts.get('Ann') and destinationID == helper.namesAndPorts.get('Chan'):    
-                PassPacket(pathAnnToChan, routerName, packetOnTheWay)
+                helper.PassPacket(pathAnnToChan, routerName, packetOnTheWay)
             
             # Packet is from Jan to Ann
             elif sourceID == helper.namesAndPorts.get('Chan') and destinationID == helper.namesAndPorts.get('Ann'):    
-                PassPacket(pathChanToAnn, routerName, packetOnTheWay)
+                helper.PassPacket(pathChanToAnn, routerName, packetOnTheWay)
 
             # Packet has no right direction
             else:
